fix overwrite filter without trigger previous fitlter removed notification (#9)
* (fix) create filter

* (feat) use zoned date time

* (feat) test listen filter overwrite

* (fix) test overwrite expired filter

+++ filter-service-core/src/main/java/cn/leancloud/filter/service/ExpirableBloomFilterConfig.java
@@ -23,7 +23,7 @@ final class ExpirableBloomFilterConfig extends AbstractBloomFilterConfig<Expirab
         return validPeriod;
     }
 
-    ExpirableBloomFilterConfig setValidPeriod(long validPeriod) {
+    ExpirableBloomFilterConfig setValidPeriod(int validPeriod) {
         checkParameter("validPeriod",
                 validPeriod >= 0,
                 "expected: >= 0, actual: %s",
